utils/config.py
import yaml


class Config(object):
    def __init__(self, config_path):
        def __dict2attr(d, prefix=''):
            for k, v in d.items():
                if isinstance(v, dict):
                    __dict2attr(v, f'{prefix}{k}_')
                else:
                    self.__setattr__(f'{prefix}{k}', v)
        with open('./config/base.yaml') as file:  # load base yaml
            default_config_dict = yaml.load(file, Loader=yaml.FullLoader)
        with open(config_path) as file:
            config_dict = yaml.load(file, Loader=yaml.FullLoader)
        __dict2attr(default_config_dict)
        __dict2attr(config_dict)

    def __getattr__(self, item):
        attr = self.__dict__.get(item)
        if attr is None:
            attr = dict()
            prefix = f'{item}_'
            for k, v in self.__dict__.items():
                if k.startswith(prefix):
                    n = k.replace(prefix, '')
                    attr[n] = v
            return attr if len(attr) > 0 else None
        else:
            return attr
